i have no idea

+++ src/main/java/data/SlayRoboto.java
@@ -9,18 +9,19 @@ import javax.persistence.Id;
 public class slayroboto {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
+//	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	
 	private int id;
 	private int base_speed;
 	private int cycle;
-	private int safety_distance;
+	private float safety_distance;
 	private int linecolor;
 	
 	public slayroboto() {
 		super();
 	}
 	
-	public slayroboto(int id, int base_speed, int cycle, int safety_distance, int linecolor) {
+	public slayroboto(int id, int base_speed, int cycle, float safety_distance, int linecolor) {
 		super();
 		this.id = 1;
 		this.base_speed = base_speed;
@@ -49,7 +50,7 @@ public class slayroboto {
 		return safety_distance;
 	}
 
-	public void setSafety_distance(int safety_distance) {
+	public void setSafety_distance(float safety_distance) {
 		this.safety_distance = safety_distance;
 	}
 
@@ -62,6 +63,10 @@ public class slayroboto {
 	}
 	
 	//For printing out the values
+	
+//		public String toString() {
+//			return "Base Speed: " + this.base_speed + "Cycle: " + this.cycle + "safety_distance: " + this.safety_distance + "Linecolor: " + this.linecolor;
+//		}
 
 	public String toString() {
 		return  this.base_speed+ " " + this.cycle + " " + this.safety_distance + " " + this.linecolor;
